// packages/whatsapp-deno-server/webhooks.types.ts
export interface Error {
  /**
   * Error code.
   *
   * We recommend that you build your app's error handling around error codes
   * instead of subcodes or HTTP response status codes.
   *
   * For details see https://developers.facebook.com/docs/whatsapp/cloud-api/support/error-codes#error-codes
   */
  code: number;
  /** Error code title. Example: Rate limit hit. */
  title: string;
  /** Error code message. This value is the same as the title value. */
  message: string;
  /** An error data object */
  error_data: ErrorDataObject;
}

export interface ErrorDataObject {
  /** Describes the error. Example: Message failed to send */
  details: string;
}
// ...
export interface Status {
  /** Information about the conversation. */
  conversation: Conversation;
  /** An array of error objects describing the error. */
  errors?: Error[];
  /** The ID for the message that the business sent to a customer */
  id: string;
  /** An object containing billing information. */
  pricing: Pricing;
  /** The WhatsApp ID for the customer */
  recipient_id: string;
  status: StatusType;
  /** Date for the status message */
  timestamp: string;
}

export interface Conversation {
  /**
   * Represents the ID of the conversation the given status notification
   * belongs to.
   */
  id: string;

  /** Indicates who initiated the conversation */
  origin: {
    /**
     * Indicates where a conversation has started.
     *
     * This can also be referred to as a conversation entry point
     */
    type: ConversationType;
  };

  /**
   * Date when the conversation expires.
   *
   * This field is only present for messages with a `status` set to `sent`.
   */
  expiration_timestamp: string;
}

export type ConversationType =
  | "business_initiated"
  | "customer_initiated"
  | "referral_conversion"
  | "user_initiated";

export interface Pricing {
  /** Indicates whether the conversation is billable */
  billable: boolean;
  /** Indicates the conversation pricing category */
  category: ConversationType;
  /**
   * Type of pricing model used.
   *
   * Current supported value is CBP
   */
  pricing_model: "CBP";
}

export type StatusType =
  | "delivered"
  | "failed"
  | "read"
  | "sent";
// ...
export type StatusListener = (
  status: Status,
) => void | Promise<void>;